+++ lib/src/context/platform_info/platform_info_io.dart
@@ -1,0 +1,12 @@
+import 'dart:io' as io;
+
+import 'platform_info.dart';
+
+PlatformInfo loadPlatformInfo() {
+  return PlatformInfo(
+    operatingSystem: io.Platform.operatingSystem,
+    operatingSystemVersion: io.Platform.operatingSystemVersion,
+    userAgent: null,
+    browserName: null,
+  );
+}