+++ lib/swetrix_flutter.dart
@@ -1,0 +1,11 @@
+export 'src/context/context.dart';
+export 'src/error_event.dart';
+export 'src/options.dart';
+export 'src/performance_metrics.dart';
+export 'src/request_options.dart';
+export 'src/swetrix.dart';
+export 'src/exceptions/swetrix_exception.dart';
+
+export 'src/context/flutter_context_builder.dart' show SwetrixFlutterContextBuilder, SwetrixFlutterEnvironment;
+export 'src/swetrix_flutter_client.dart' show SwetrixFlutterClient, SharedPreferencesFactory;
+export 'src/context/visitor/visitor_store.dart' show SwetrixVisitorStore;